+++ CMakeLists.txt
@@ -291,6 +291,7 @@ set(SOURCES
         db/compaction_iterator.cc
         db/compaction_job.cc
         db/compaction_picker.cc
+        db/compaction_picker_universal.cc
         db/convenience.cc
         db/db_filesnapshot.cc
         db/db_impl.cc
